codereview fixes
